Refactor option classes to their own file.
